app/backend/pyproject.toml
[tool.poetry]
name = "bolao-backend"
version = "0.1.0"
description = "Backend API do Bolão Brasileirão"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = { extras = ["standard"], version = "^0.34" }
sqlalchemy = "^2.0"
pydantic-settings = "^2.0"
alembic = "^1.14"
httpx = "^0.28"
apscheduler = "^3.10"
cloudscraper = "^1.2.71"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = "^1.7.4"
bcrypt = "^5.0.0"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
